requirements.txt
beartype
colorama
google-generativeai
google.ai.generativelanguage
langchain
mistralai
pydantic
pytest
pyyaml
rich
